refactor to externalize scripts
