+++ package.json
@@ -1,18 +1,17 @@
 {
   "name": "svg-gobbler",
-  "private": true,
-  "version": "5.2.0",
+  "version": "5.3.0",
   "description": "Organize, optimize, download, copy, edit, and export SVGs.",
   "author": "Ross Moody<ross@rossmoody.com>",
   "license": "apache-2.0",
   "type": "module",
-  "homepage": "https://www.svggobbler.com",
+  "module": "dist/gather-page-data.js",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "serve": "nodemon server/index.ts --exec 'ts-node --esm'",
     "clean": "rm -rf dist",
-    "release": "yarn clean && yarn build && npx ts-node --esm src/utils/zip-release.ts"
+    "release": "yarn build && npx ts-node --esm src/utils/zip-release.ts"
   },
   "dependencies": {
     "@codemirror/lang-html": "^6.4.6",
@@ -34,6 +33,7 @@
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-router-dom": "^6.18.0",
+    "svg-gobbler-scripts": "^1.0.6",
     "svgo": "^3.0.3"
   },
   "devDependencies": {
@@ -49,7 +49,7 @@
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "@uiw/codemirror-theme-tokyo-night-storm": "^4.21.20",
-    "@vitejs/plugin-react-swc": "^3.3.2",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "archiver": "^6.0.1",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.45.0",
@@ -62,6 +62,6 @@
     "tailwindcss": "^3.3.5",
     "tailwindcss-radix": "^2.8.0",
     "typescript": "^5.0.2",
-    "vite": "^5.0.0"
+    "vite": "^5.0.5"
   }
 }
