+++ src/layout/collection/card-context-menu/index.tsx
@@ -6,7 +6,7 @@ import {
 } from '@heroicons/react/24/outline'
 import * as ContextMenu from '@radix-ui/react-context-menu'
 import { Fragment, PropsWithChildren } from 'react'
-import type { Svg } from 'scripts/svg-classes/svg'
+import type { Svg } from 'svg-gobbler-scripts'
 import { ContextMenuItem } from './context-menu-item'
 import { ContextMenuSubTrigger } from './context-menu-sub-trigger'
 import { useContextActions } from './use-context-actions'
