+++ src/layout/collection/card/cors-restricted-actions.tsx
@@ -1,6 +1,6 @@
 import { ArrowTopRightOnSquareIcon, InformationCircleIcon } from '@heroicons/react/24/outline'
-import { Image } from 'scripts/svg-classes/image'
 import { Button, Tooltip } from 'src/components'
+import type { Image } from 'svg-gobbler-scripts'
 
 type Props = {
   data: Image
