+++ src/layout/collection/card/index.tsx
@@ -1,8 +1,7 @@
 import clsx from 'clsx'
 import { HTMLAttributes, forwardRef, useMemo } from 'react'
-import type { Image } from 'scripts/svg-classes/image'
-import type { Svg } from 'scripts/svg-classes/svg'
 import { useCollection } from 'src/providers'
+import type { Image, Svg } from 'svg-gobbler-scripts'
 import { CardContextMenu } from '../card-context-menu'
 import { CardContent } from './card-content'
 import { CorsRestrictedActions } from './cors-restricted-actions'
