+++ src/types.ts
@@ -1,8 +1,4 @@
-import type { GElement } from 'scripts/svg-classes/g-element'
-import type { Image } from 'scripts/svg-classes/image'
-import type { Inline } from 'scripts/svg-classes/inline'
-import type { Svg } from 'scripts/svg-classes/svg'
-import type { SvgSymbol } from 'scripts/svg-classes/symbol'
+import type { GElement, Image, Inline, Svg, SvgSymbol } from 'svg-gobbler-scripts'
 
 /**
  * The message sent from the Background script to the Content script.
