+++ tsconfig.json
@@ -8,7 +8,6 @@
     "baseUrl": ".",
     "paths": {
       "src/*": ["src/*"],
-      "scripts/*": ["scripts/*"],
       "data/*": ["data/*"],
       "types": ["src/types.ts"]
     },
@@ -27,6 +26,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src", "scripts"],
+  "include": ["src", "scripts", "background.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
